feat: totalKilo Double로 변경

+++ core/src/main/kotlin/tamago/server/core/user/domain/aggregate/UserRunningData.kt
@@ -2,13 +2,13 @@ package tamago.server.core.user.domain.aggregate
 
 data class UserRunningData(
     val goalKilo: Int? = null,
-    val totalKilo: Int? = null,
+    val totalKilo: Double? = null,
 ) {
     fun updateGoalKilo(goalKilo: Int): UserRunningData {
         return copy(goalKilo = goalKilo)
     }
 
-    fun addKilo(kilo: Int): UserRunningData {
-        return copy(totalKilo = (totalKilo ?: 0) + kilo)
+    fun addKilo(kilo: Double): UserRunningData {
+        return copy(totalKilo = (totalKilo ?: 0.0) + kilo)
     }
 }
