+++ core/src/main/kotlin/tamago/server/core/user/infrastructure/entity/UserEntity.kt
@@ -33,6 +33,6 @@ class UserEntity(
     )
     val auths: MutableList<UserAuthEntity> = mutableListOf(),
     var goalKilo: Int? = null,
-    var totalKilo: Int? = null,
+    var totalKilo: Double? = null,
     var lastLoginAt: LocalDateTime? = null,
 ) : BaseTimeEntity()
